Rename contracts, remove dupes
