+++ libs/src/services/contracts/GovernedContractClient.ts
@@ -1,5 +1,5 @@
 import type { ContractABI, Logger } from '../../utils'
-import type { GovernanceClient } from '../ethContracts/governanceClient'
+import type { GovernanceClient } from '../ethContracts/GovernanceClient'
 import type { EthWeb3Manager } from '../ethWeb3Manager'
 import type { Web3Manager } from '../web3Manager'
 import type { GetRegistryAddress } from './ContractClient'
