+++ libs/src/services/ethContracts/GovernanceClient.ts
@@ -1,8 +1,8 @@
 import { ContractClient, GetRegistryAddress } from '../contracts/ContractClient'
 import { ContractABI, ContractMethod, Logger, Utils } from '../../utils'
 import type { EthWeb3Manager } from '../ethWeb3Manager'
-import type { AudiusTokenClient } from './audiusTokenClient'
-import type { StakingProxyClient } from './stakingProxyClient'
+import type { AudiusTokenClient } from './AudiusTokenClient'
+import type { StakingProxyClient } from './StakingProxyClient'
 import type { EventLog } from 'web3-core'
 
 type ProposalTxn = {
