+++ libs/src/services/ethContracts/StakingProxyClient.ts
@@ -1,7 +1,7 @@
 import type { ContractABI, Logger } from '../../utils'
 import { ContractClient, GetRegistryAddress } from '../contracts/ContractClient'
 import type { EthWeb3Manager } from '../ethWeb3Manager'
-import type { AudiusTokenClient } from './audiusTokenClient'
+import type { AudiusTokenClient } from './AudiusTokenClient'
 import type BN from 'bn.js'
 
 export class StakingProxyClient extends ContractClient {
