+++ libs/src/services/ethContracts/wormholeClient.ts
@@ -1,7 +1,7 @@
 import type Web3 from 'web3'
 import type { ContractABI } from '../../utils'
 import type { EthWeb3Manager } from '../ethWeb3Manager'
-import type { AudiusTokenClient } from './audiusTokenClient'
+import type { AudiusTokenClient } from './AudiusTokenClient'
 import type { Contract } from 'web3-eth-contract'
 import type Wallet from 'ethereumjs-wallet'
 import type BN from 'bn.js'
